requirements.txt
streamlit 
PyMuPDF 
nltk 
sentence-transformers 
faiss-cpu 
requests 
scikit-learn 
python-dotenv
langchain-groq